n == 0 should imply infinite criteria value
This commit fixes https://github.com/JuliaPOMDP/POMCPOW.jl/issues/30

If the value for this node/action is already an infinity,
then we keep that infinity value as the criterion value, since UCB cannot change it.
At best, if the value were -infinity and n==0, we get -Inf + Inf == NaN, which would not be good.

Prior to this fix, if node values are initialized to zero and rewards are positive,
sometimes the first node would be explored twice before the second node gets explored the first time.